Factoring out GW config into separate file
