+++ lib/holiday-stops/src/test/scala/com/gu/holiday_stops/HolidayCreditSpec.scala
@@ -28,13 +28,13 @@ object HolidayCreditSpec extends Properties("HolidayCreditAmount") with OptionVa
   val subscription = Fixtures.mkSubscription()
 
   property("should be negative") = forAll(ratePlanChargeGen) { charge: RatePlanCharge =>
-    val ratePlans = List(RatePlan("", List(charge), Config.guardianWeeklyProductRatePlanIdsPROD.head, ""))
+    val ratePlans = List(RatePlan("", List(charge), GuardianWeeklyHolidayStopConfig.guardianWeeklyProductRatePlanIdsPROD.head, ""))
     val currentGuardianWeeklySubscription = CurrentGuardianWeeklySubscription(subscription.copy(ratePlans = ratePlans), Fixtures.guardianWeeklyConfig.guardianWeeklyProductRatePlanIds, Nil)
     HolidayCredit(currentGuardianWeeklySubscription.right.value, LocalDate.now) < 0
   }
 
   property("should never be overwhelmingly negative") = forAll(ratePlanChargeGen) { charge: RatePlanCharge =>
-    val ratePlans = List(RatePlan("", List(charge), Config.guardianWeeklyProductRatePlanIdsPROD.head, ""))
+    val ratePlans = List(RatePlan("", List(charge), GuardianWeeklyHolidayStopConfig.guardianWeeklyProductRatePlanIdsPROD.head, ""))
     val currentGuardianWeeklySubscription = CurrentGuardianWeeklySubscription(subscription.copy(ratePlans = ratePlans), Fixtures.guardianWeeklyConfig.guardianWeeklyProductRatePlanIds, Nil)
     HolidayCredit(currentGuardianWeeklySubscription.right.value, LocalDate.now) > -charge.price
   }
